Lab_01_07_04 updated

+++ lab_01_07_04/func_tests/scripts/pos_case.sh
@@ -6,17 +6,9 @@ if [ -z "$1" ] || [ -z "$2" ] ; then
 fi
 
 file="$(mktemp)"
-../../app.exe < "$1" > "$file"
+./app.exe < "$1" > "$file"
 
-if [ -z "$(cat "$2")" ]; then
-  if [ "$(cat "$file")" == "Enter number to work with: " ]; then
-    exit 0
-  else
-    exit 1
-  fi
-fi
-
-if ./comparator.sh "$file" "$2"; then
+if func_tests/scripts/comparator.sh "$file" "$2"; then
   rm "$file"
   exit 0
 else
